Lexer: Define the basic `Preprocessing` module.
