+++ tests/samples/lexer/boolean_literal.cc
@@ -1,0 +1,5 @@
+int main(){
+    true;
+    false;
+    nullptr;
+}
